azure_get_user_group_membership_and_last_signin/pyproject.toml
[tool.poetry]
name = "azure-get-user-group-membership-and-last-signin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
msgraph-sdk = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
